# historian/backend/src/main/swagger/api-swagger.yml
swagger: '2.0'
info:
  description: REST API for logisland historian
  version: v1
  title: Logisland Data Historian
  contact:
    name: hurence
    email: support@hurence.com
host: localhost:8701
basePath: /api/v1
schemes:
  - http
  - https
paths:
...
  /selections:
      get:
        tags:
          - selection
        operationId: getAllUserSelection
        summary:  get all slections that this user has access to
        description: retrieve all slections that user has access to
        produces:
          - application/json
        responses:
          '200':
             description: Selection list
             schema:
               type: array
               items:
                 $ref: '#/definitions/Selection'
          default:
             description: unexpected error
             schema:
               $ref: '#/definitions/Error'
  /selections/{selectionId}:
      get:
       tags:
         - selection
       summary: get Selection
       description: get the corresponding Selection
       operationId: getSelection
       produces:
         - application/json
       parameters:
         - name: selectionId
           in: path
           description: id of the Selection to return
           required: true
           type: string
       responses:
         '200':
           description: Selection found
           schema:
             $ref: '#/definitions/Selection'
         '404':
            description: Selection resource not found
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
      delete:
       tags:
         - selection
       summary: delete Selection
       description: remove the corresponding Selection
       operationId: deleteSelection
       parameters:
         - name: selectionId
           in: path
           description: id of the Selection to be deleted
           required: true
           type: string
       responses:
         '404':
           description: Selection resource not found
         '400':
           description: Invalid ID supplied
         '200':
           description: Selection successfully removed
           schema:
             $ref: '#/definitions/Selection'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
      put:
       tags:
        - selection
       summary:  update Selection
       description: update an existing Selection
       operationId: updateSelection
       parameters:
         - name: selectionId
           in: path
           description: selectionId to be updated
           required: true
           type: string
         - name: body
           in: body
           description: new Selection definition
           required: true
           schema:
             $ref: '#/definitions/Selection'
       responses:
        '404':
          description: Selection resource not found
        '400':
          description: Invalid ID supplied
        '200':
         description: Selection successfuly updated
         schema:
           $ref: '#/definitions/Selection'
      post:
       tags:
        - selection
       summary:  create new Selection
       description:  store a new Selection
       operationId: addSelectionWithId
       parameters:
         - name: body
           in: body
           description: Selection resource to add
           required: true
           schema:
             $ref: '#/definitions/Selection'
         - name: selectionId
           in: path
           description: selectionId to create
           required: true
           type: string
       responses:
         '400':
           description: Invalid ID supplied
         '200':
           description: Selection successfuly created
           schema:
             $ref: '#/definitions/Selection'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
  /selections/{selectionId}/tags:
       get:
         tags:
          - selection
          - tag
         operationId: getAllTagsFromSelection
         summary:  get all tags from a selection
         description: retrieve all tags through this selection
         produces:
            - application/json
         parameters:
           - name: selectionId
             in: path
             description: id of the selection
             required: true
             type: string
         responses:
           '200':
             description: Tag list
             schema:
               type: array
               items:
                 $ref: '#/definitions/Tag'
           default:
             description: unexpected error
             schema:
               $ref: '#/definitions/Error'
...
definitions:
...
  Selection:
    type: object
    description: a Selection is a selection of tags.
    required:
      - id
      - tags
    properties:
      id:
        description: concatenation of owner with name for example
        type: string
      name:
        type: string
      owner:
        type: string
      tags:
        description: list of tags id in selection
        type: array
        items:
          type: string
      record_type:
        type: string
        default: 'selection'
  Tag:
    type: object
    description: a Tag is an identifier to an OPC value
    required:
      - item_name
      - datasource_id
      - data_type
    properties:
      record_type:
        type: string
        default: 'tag'
      id:
        type: string
        default: 'mySweetUniqueId'
      datasource_id:
        type: string
      domain:
        type: string
      server:
        type: string
      group:
        type: string
      tag_name:
        type: string
      labels:
        description: can be used to group tags around a theme
        type: array
        items:
          type: string
      data_type:
        type: string
        default: float
        enum:
        - int
        - long
        - float
        - double
        - string
        - array
        - null
        - bytes
        - boolean
      description:
        type: string
      text:
        description: ctach all field
        type: array
        items:
          type: string
      creation_date:
        description: timestamp
        type: integer
        format: int64
      last_modification_date:
        description: timestamp
        type: integer
        format: int64
      last_polling_date:
        description: timestamp
        type: integer
        format: int64
      update_rate:
        description: polling delay in ms
        default: 1000
        type: integer
      min_numeric_value:
        type: number
        format: dumble
      max_numeric_value:
        type: number
        format: double
      last_numeric_value:
        type: number
        format: double
      last_quality:
        type: integer
...
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string